test: 의미없는 테스트 삭제 및 클래스에 맞춰서 테스트 이동
